Moved files to common directory.
